Working on unit testing
