+++ test/fc_base/test_state_field.cpp
@@ -4,12 +4,24 @@
 #include <unity.h>
 
 void test_internal_state_field() {
+    // Test setter and getter for state field.
+
     InternalStateField<float> isf("field");
     TEST_ASSERT_FALSE(isf.is_readable());
     TEST_ASSERT_FALSE(isf.is_writable());
 
-    isf.set(2);
+    isf.set(2.0f);
     TEST_ASSERT_EQUAL(isf.get(), 2);
+}
+
+void test_readable_state_field() {
+    Serializer<float> serializer(0, 10, 10);
+    ReadableStateField<float> rsf("field", serializer);
+    TEST_ASSERT(rsf.is_readable());
+    TEST_ASSERT_FALSE(rsf.is_writable());
+
+    rsf.set(2.0f);
+    TEST_ASSERT_EQUAL(rsf.get(), 2);
 }
 
 void test_state_field() { RUN_TEST(test_internal_state_field); }
